- added cflags (Wno-misleading-indentation, Wno-ignored-attributes) to makefile generation (premake4.lua) -> needed to compile with gcc-6.6 - fixed some conversion errors

+++ NaoTHSoccer/Source/Cognition/Modules/Modeling/RoleDecision/StableRoleDecision/StableRoleDecision.cpp
@@ -56,7 +56,7 @@ void StableRoleDecision::execute()
     }
 
     double time_bonus = 0;
-    if(msg.playerNum==getRoleDecisionModel().firstStriker) {
+    if(msg.playerNum==(unsigned int)getRoleDecisionModel().firstStriker) {
       time_bonus = parameters.strikerBonusTime;
     }
